+++ src/Controller/SamlSsoController.php
@@ -47,7 +47,6 @@ namespace GlpiPlugin\Samlsso\Controller;
 use Glpi\Http\Firewall;                                                 // Required to allow anonymous access to ACS route
 use Glpi\Controller\AbstractController;                                 // The controller 
 use Glpi\Security\Attribute\SecurityStrategy;                           // Required to decorate the invoke
-use Glpi\Security\Attribute\DisableCsrfChecks;                          // Required to disable the Csrf checks
 use Symfony\Component\HttpFoundation\Request;                           // Required for __invoke
 use Symfony\Component\HttpFoundation\Response;                          // Required for __invoke
 use Symfony\Component\Routing\Attribute\Route;                          // Required to register controller route
@@ -56,36 +55,30 @@ use GlpiPlugin\Samlsso\LoginFlow\Meta;                                  // Requi
 use GlpiPlugin\Samlsso\Config\ConfigForm;                               // Required to call Config object
 use GlpiPlugin\Samlsso\Exclude;                                         // Required to call Exclude object
 use GlpiPlugin\Samlsso\RuleSaml;                                        // Required to call Rules object
-use GlpiPlugin\Samlsso\LoginState;
-use GlpiPlugin\Samlsso\LoginFlow\LoginFlowForm;
-
-
-
+use GlpiPlugin\Samlsso\LoginState;                                      //
+use GlpiPlugin\Samlsso\LoginFlow\LoginFlowForm;                         //
 
 final class SamlSsoController extends AbstractController
 {
+    ####################################################################
     // ACS route
     public const ACS_ROUTE      = 'front/acs';                          // Route being registered by __class__
     public const ACS_PARAM      = '/{'.LoginState::IDP_ID.'}';
     public const ACS_NAME       = 'samlsso_ACS';                        // Route name
 
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
-    #[Route(self::ACS_ROUTE.self::ACS_PARAM, name: self::ACS_NAME)]                     // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
+    #[Route(self::ACS_ROUTE.self::ACS_PARAM, name: self::ACS_NAME)]     // Decorator to register route to controller
     public function acs(Request $request): Response                     // What to do if route is invoked.
     {
-        return new Response((new Acs)->init($request));            // Call the ACS handler.
+        return new Response((new Acs)->init($request));                 // Call the ACS handler.
     }
 
 
-
+    ####################################################################
     // SLO route
     public const SLO_ROUTE      = 'front/slo';                          // Route being registered by __class__
     public const SLO_NAME       = 'samlsso_SLO';                        // Route name
- 
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
-    #[Route(self::SLO_ROUTE, name: self::SLO_NAME)]                     // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function slo(Request $request): Response                     // What to do if route is invoked.
     {
         global $CFG_GLPI;
@@ -93,7 +86,7 @@ final class SamlSsoController extends AbstractController
     }
 
 
-
+    ####################################################################
     // Meta route
     public const META_ROUTE     = 'front/meta';                         // Route being registered by __class__
     public const META_PARAM     = '/{idpId}';
@@ -101,43 +94,42 @@ final class SamlSsoController extends AbstractController
 
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
     #[Route(self::META_ROUTE.self::META_PARAM, name: self::META_NAME)]  // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function meta(Request $request): Response                    // What to do if route is invoked.
     {
         return (new Meta)->getSPMeta($request);                         // Call the SPMeta handler.
     }
 
 
-
+    ####################################################################
     // Config routes
     public const CONFIG_FILE     = 'front/config.php';                  // Register old route as well
     public const CONFIG_ROUTE    = 'front/config';                      // Route being registered by __class__
     public const CONFIG_NAME     = 'configMain';                        // Route name
     public const CONFIG_PNAME    = 'config';                            // Parent object name
-    public const CONFIGFORM_FILE = 'front/config.form.php';             // Register old route as well
-    public const CONFIGFORM_ROUTE= 'front/configform/';                 // Route being registered by __class__
-    public const CONFIGFORM_NAME = 'configForm';                        // Route name
-    public const CONFIGFORM_PNAME= 'config';                            // Parent object name
 
-    #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
+    //#[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                  // Decorator to disable authentication check
     #[Route(self::CONFIG_ROUTE, name: self::CONFIG_NAME)]               // Decorator to register route to controller
     #[Route(self::CONFIG_FILE, name: self::CONFIG_NAME.'_file')]        // Decorator to register old route to handle GLPI generated menu's
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function config(Request $request): Response                  // What to do if route is invoked.
     {
-        return new Response((new ConfigForm)->invoke($request));        // Call the ACS handler.
+        return new Response((new ConfigForm)->invoke($request));        // Call the form handler.
     }
-    #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
+
+    // ConfigForm routes
+    public const CONFIGFORM_FILE = 'front/config.form.php';             // Register old route as well
+    public const CONFIGFORM_ROUTE= 'front/config/form';                 // Route being registered by __class__
+    public const CONFIGFORM_NAME = 'configForm';                        // Route name
+    public const CONFIGFORM_PNAME= 'config';                            // Parent object name
+    //#[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                  // Decorator to disable authentication check
     #[Route(self::CONFIGFORM_ROUTE, name: self::CONFIGFORM_NAME)]       // Decorator to register route to controller
     #[Route(self::CONFIGFORM_FILE, name: self::CONFIGFORM_NAME.'_file')]// Decorator to register old route to handle GLPI generated menu's
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function configform(Request $request): Response
     {
-        return new Response((new ConfigForm)->invokeForm($request));    // Call the ACS handler.
+        return new Response((new ConfigForm)->invokeForm($request));    // Call the form handler.
     }
 
 
-
+    ####################################################################
     // LoginFlowConfig
     public const FLOWFORM_FILE = 'front/loginflow.form.php';
     public const FLOWFORM_ROUTE= 'front/flowconfig';                    // Route being registered by __class__
@@ -146,29 +138,26 @@ final class SamlSsoController extends AbstractController
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
     #[Route(self::FLOWFORM_ROUTE, name: self::FLOWFORM_NAME)]           // Decorator to register route to controller
     #[Route(self::FLOWFORM_FILE, name: self::FLOWFORM_NAME.'_file')]    // Decorator to register old route to handle GLPI generated menu's
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function loginflow(Request $request): Response               // What to do if route is invoked.
     {
         return new Response((new LoginFlowForm)->init($request));       // Call the ACS handler.
     }
 
 
-
+    ####################################################################
     // Exclude routes
     public const EXCLUDE_ROUTE = 'front/exclude';                       // Route being registered by __class__
     public const EXCLUDE_NAME  = 'excludeMain';                         // Route name
     public const EXCLUDE_PNAME = 'config';                              // Parent object name
-
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
     #[Route(self::EXCLUDE_ROUTE, name: self::EXCLUDE_NAME)]             // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function exclude(): Response                                 // What to do if route is invoked.
     {
         return new Response((new Exclude)->invoke());                   // Call the ACS handler.
     }
 
 
-
+    ####################################################################
     // Rules routes
     public const RULES_FILE     = 'front/rulesaml.php';                 // Register all route as well because these might be autogenerated
     public const RULES_ROUTE    = 'front/rule';                         // Route being registered by __class__
@@ -183,15 +172,14 @@ final class SamlSsoController extends AbstractController
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
     #[Route(self::RULES_ROUTE, name: self::RULES_NAME)]                 // Decorator to register route to controller
     #[Route(self::RULES_FILE, name: self::RULES_NAME.'_file')]          // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function __invoke(Request $request): Response                // What to do if route is invoked.
     {
         return new Response((new RuleSaml)->invoke());                  // Call the ACS handler.
     }
+
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]                    // Decorator to disable authentication check
     #[Route(self::RULESFORM_ROUTE, name: self::RULESFORM_NAME)]         // Decorator to register route to controller
     #[Route(self::RULESFORM_FILE, name: self::RULESFORM_NAME.'_file')]  // Decorator to register route to controller
-    #[DisableCsrfChecks()]                                              // Decorator to disable Csrf checking
     public function itemform(Request $request): Response                // What to do if route is invoked.
     {
         return new Response((new RuleSaml)->invokeForm());              // Call the ACS handler.